+++ app/src/main/java/ucoc/weatherapp/Models/Wind.java
@@ -2,7 +2,7 @@ package ucoc.weatherapp.Models;
 
 public class Wind {
     private double speed;
-    private int deg;
+    private float deg;
 
     public Wind() {
     }
@@ -15,11 +15,13 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
 
-    public void setDeg(int deg) {
+    public void setDeg(float deg) {
         this.deg = deg;
     }
 }
+
+
